Do not steal libbpf's file descriptors
This is for I/O safety: We currently create some mmap on a file
descriptor that we do *not* own - its owned by libbpf. Generally,
the FD lifetime is bound by the Map's lifetime. However, our RingBuffer
struct is not bound to the Map's lifetime, and may well outlive it. This
is unsafe, because if the Map gets dropped, the FD is lost and then we
rely on things that may no longer be there.

This change fixes this, by duplicating the file descriptor. Thus, we
actually own it and can therefore be sure of its lifetime. This makes
our RingBuffer truly independent of the original Maps reference
lifetime, as implied by the function contract of RingBuffer::new().